backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.30
alembic>=1.13.2
psycopg[binary]>=3.2.1
pydantic-settings>=2.2.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=3.2.0,<4.1
python-multipart>=0.0.9
httpx>=0.27.0
requests>=2.31.0
prometheus-client>=0.20.0
paho-mqtt>=2.1.0
